# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Travel App - приложение для путешественников",
        "title": "Prisma",
        "contact": {},
        "version": "1.0"
    },
    "host": "prisma.ssapi.ru",
    "basePath": "/",
    "paths": {
...
        "/users/me": {
            "get": {
                "description": "Возвращает информацию о текущем пользователе по его VK ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить информацию о текущем пользователе",
                "operationId": "get-me",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.UserGetMeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "description": "Редактирует профиль текущего пользователя.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Редактировать профиль пользователя",
                "operationId": "edit-user",
                "parameters": [
                    {
                        "description": "Пройдено обучение (опционально)",
                        "name": "passed_onboarding",
                        "in": "body",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "description": "Широта выбранного местоположения",
                        "name": "selected_geo_lat",
                        "in": "body",
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "description": "Долгота выбранного местоположения",
                        "name": "selected_geo_lot",
                        "in": "body",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/{vkId}": {
            "get": {
                "description": "Возвращает информацию о пользователе по его VK ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Получить пользователя по VK ID",
                "operationId": "get-user-by-vk-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Уникальный идентификатор пользователя в VK",
                        "name": "vkId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.User"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {}
            }
        },
...
        "models.User": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "passed_onboarding": {
                    "type": "boolean"
                },
                "selected_geo": {
                    "type": "string"
                },
                "vk_id": {
                    "type": "integer"
                }
            }
        },
        "models.UserGetMeResponse": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "current_geo": {
                    "type": "string"
                },
                "is_admin": {
                    "type": "boolean"
                },
                "passed_onboarding": {
                    "type": "boolean"
                },
                "selected_geo": {
                    "type": "string"
                },
                "vk_id": {
                    "type": "integer"
                }
            }
        }
    }
}